requirements.txt
fastapi==0.70.0
numpy==1.19.5
statistics==1.0.3.5
thefuzz==0.19.0
jellyfish==0.8.9
uvicorn==0.15.0
PyJWT==2.3.0
python-multipart==0.0.5
python-jose[cryptography]
passlib[bcrypt]